Up to date with origin

+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/policy/linked/LRBBBlock.java
@@ -149,7 +149,7 @@ public final class LRBBBlock {
   private Node addToList(AccessEvent candidate, Node inSentinel) {
     Node node = new Node(candidate, candidate.weight(), inSentinel);
 
-    data.put((long) candidate.key(), node);
+    data.put(candidate.key(), node);
     if (inSentinel.size > 0) {
       Node listNext = inSentinel.next;
       if (listNext.event.delta() * Math.pow((double) currOp - listNext.lastOp + 1, -k) > candidate
